genetic algorithm working with strings as population members

+++ parallel_computing/Person.java
@@ -1,14 +1,16 @@
+
 class Person {
 
 	String DNA;
-	float fitness;
+	int fitness;
 	
 	Person(String DNA){
 		this.DNA = DNA;
 	}
 
 	
-	public void set_fitness(float fitness){
+	public void set_fitness(int fitness){
 		this.fitness = fitness;
 	}
-}	
+
+}
